sensor/requirements.txt
RPi.bme280
smbus2
paho-mqtt
pyyaml